[windows] Always use system-shipped tar (#1)
On some Windows runners, it is possible for a GNU tar binary to be
further ahead in `PATH`. In particular, Git for Windows ships with a GNU
tar binary, which can't parse Windows-style paths properly. This works
around the issue by always using the BSD tar binary that always ships
with the operating system on Windows.

Co-authored-by: Saleem Abdulrasool <compnerd@compnerd.org>